requirements.txt
streamlit
langchain
langchain-openai
tiktoken
chromadb
openai
python-dotenv
pysqlite-binary
PyPDF2
